module to aid against csrf attacks
The module implements a cookie-to-header protection mechanism by sending
out a token as a cookie and by expecting the data of the cookie in a
header of the incoming request.

The validation of the token only takes place for configured "protected" URLs,
so an "entry" to a page (i.e. index.html) can be unprotected and thus be
used for obtaining a cookie.

The token is based on an sha1 hmac where the input data are the
timestamp and the username (if available). The secret key for the hmac
routine is generated by the SSL random function and is stored in memory.

The expiration time of the token is configurable, the token will be
updated and renewed before the expiration time elapses.

The module requires OpenSSL functions and will only be enabled if
lighttpd is compiled with OpenSSL and libcrypto support.

Configuration options are:

Name of the cookie that will be sent in outgoing responses:
csrf.send-cookie-name = "csrf";

Name of the header that will contain the token value in incoming requests:
csrf.receive-header-name = "X-Csrf-Token";

Cookie "Domain" setting, not added to the cookie if unset:
csrf.cookie-domain = "";

Cookie lifetime in seconds:
csrf.cookie-lifetime = 600;

Conditions can be used to activate protection for certain URLs:

$HTTP["url"] =~ "^/(someurl|cgi-bin)/(.+)" {
    csrf.protection = "enable"
}